Laravel Admin is Done

+++ laravel-admin/app/Http/Controllers/UserController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\UserUpdateRequest;
 use App\Http\Resources\UserResource;
 use App\Models\User;
 use Auth;
+use Gate;
 use Illuminate\Http\Request;
 use Laravel\Passport\RefreshToken;
 use Laravel\Passport\Token;
@@ -18,16 +19,19 @@ class UserController extends Controller
 {
     public function index()
     {
+        Gate::authorize('view', 'users');
         $users = User::paginate();
         return UserResource::collection($users);
     }
     public function show($id)
     {
+        Gate::authorize('view', 'users');
         $user = User::findOrFail($id);
         return new UserResource($user);
     }
     public function store(UserCreateRequest $request)
     {
+        Gate::authorize('edit', 'users');
         $email = $request->input('email');
         if (User::where('email', $email)->exists()) {
             return response()->json(['message' => 'User already exists'], Response::HTTP_CONFLICT);
@@ -39,18 +43,25 @@ class UserController extends Controller
     }
     public function update(UserUpdateRequest $request, $id)
     {
+        Gate::authorize('edit', 'users');
         $user = User::findOrFail($id);
         $user->update($request->only(['first_name', 'last_name', 'email','role_id']));
         return response()->json(new UserResource($user), Response::HTTP_ACCEPTED);
     }
     public function destroy($id)
     {
+        Gate::authorize('edit', 'users');
         User::findOrFail($id)->delete();
         return response('Deleted Successfully', Response::HTTP_NO_CONTENT);
     }
     public function user()
     {
-        return new UserResource(auth()->user());
+        $user = Auth::user();
+        return (new UserResource($user))->additional([
+            'data' => [
+              'permissions' => $user->permissions(),
+            ],
+        ]);
     }
     public function updateInfo(UserUpdateRequest $request)
     {
